pyproject.toml
[tool.poetry]
name = "gitlab-due-sync"
version = "0.1.0"
description = ""
authors = ["Sarisia <sarisiaice@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"

[tool.poetry.dev-dependencies]
aws-cdk-lib = "^2.25.0"
constructs = "^10.1.22"
"aws-cdk.aws-lambda-python-alpha" = "^2.26.0-alpha.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
